fix(bolt): use named imports for bbolt
The bbolt project has one version that uses the import `bolt` and one
that uses `bbolt`. The go module system doesn't really differentiate
between these two as different modules though so it is easy to switch to
using one or the other.

To remain consistent, the `bolt` named import is used everywhere so it
doesn't matter which version is used.