build(webpack): add source maps for @influxdata/giraffe prod builds
We were getting runtime errors in the prod builds via honeybadger.
They were unreadable because we did not have source maps for giraffe.

We were only able to get this working for prod builds but not dev
despite trying putting it in the webpack common file.  This is good
'nuff for our purpose now, but maybe we'll need further investigation.

Bonus commit... I added a new make target of influxd so I would not
need to type so much.

Co-authored-by: Andrew Watkins <watts@influxdb.com>