fix(onboarding): remove fancy scrollbar that hides content
e2e tests are failing because the fancyscroll bar component hides
the admin step during certain points in onboarding.

We were able to reproduce when the form in the admin step had been
marked as valid.

The use of the scrollbar in the setting seemed unneeded as it could
be rendered in a quite small window.

Therefore, we removed the scrollbar altogether for now.

Co-authored-by: Christopher Henn <chris@chrishenn.net>