fix(ui): fix max call stack exceeded error when parsing CSV
When parsing a Flux CSV file with hundreds of thousands of lines for the
"Raw Data View", we would see a

    Maximum call stack size exceeded

error. This was because every line in the CSV was being passed as an
argument in a single `Array.prototype.push` call, and there are
[limits][0] to how many arguments can be passed to a function.

This commit avoids passing the arguments all at once.

[0]: https://stackoverflow.com/questions/22747068/is-there-a-max-number-of-arguments-javascript-functions-can-accept

Closes #14566