chore(ui): relocate testing configuration (#15356)
while exploring containerization of the frontend, i decided to remove the
existing environment configuration from package.json and push them into
external files in an attempt to improve the caching strategies available
for local development. the CI pipeline should be watching the lock file,
but changes to the testing environment locally shouldn't cause a rebuild
of node_modules.