feat(verify-seriesfile): add option to filter
This commit adds an option to the `influxd inspect verify-series-file`
tool that allows an operator to only check series for a specific bucket
or all buckets in a specific org. To do this the operator provides a
base-16 encoded string representing the org_id and bucket_id or interest.

Since it's just a prefix, providing only the `org_id` would check all
buckets.