chore(pkger): improve error handling in service
this provides influxdb.Errors whenever possible from the pker service layer.
the behavior that caused the error is somewhat implicated in the errro code.
it is not perfect but is a big step forward. using the http server/client to
run pkger stuff makes it abundantly clear the errors need to be communicated
better.

closes: #16313