Add support for float64 to rocksdb
Adds support for for f64 time series in RocksDB. Series data types are now stored in the index under the id to key mapping, which is now id to type and key.

This doesn't enforce the same data type for values being written into a series, which will come later. Also later will be adding support for float64 series in the read API.