feat(models): Add limits options to line-protocol parser
The behavior of the line-protocol parser remains unchanged. This PR
moves the functions to a struct in order to simplify the tracking of
limits whilst parsing a request.

Limits may be set using the ParsePointsWithOptions API and the
associated WithParser functional options, all of which are documented.

It should be noted that the behavior of ParsePointsWithPrecision remains
unchanged.

ParsePointsWithPrecisionV1 was no longer used and removed to simplify
code.