refactor(service): start services asynchronously
These changes allow us to start services (tsdb, http, monitor,
continuous query, etc.) asynchronously so that we can access the http
server before the tsdb service is ready.  Before this patch, the http
service was loaded after tsdb which makes it impossible to retrieve
profiling information that could help diagnose issues during tsdb
startup.

This patch replaces the Open()/Close() metaphor used for starting
services with simply Run() with concurrency handled by the caller and
cancellation done via a context.  It also removes the assumption that
Run(), Open(), and Close() can be called more than once on a particular
value, and this patch provides a backwards compatible API so that we can
preserve old tests.

Following the principle of leaving concurrency to the caller, this patch
adds a Run() method which accepts a context value and blocks until it is
cancelled.  We also supply Run() with a context value to provide a
standard and idiomatic way of cancelling each service.

The cancel() function associated with a context now serves the purpose
of the Close() method -- once a service starts, it blocks on ctx.Done()
and performs any necessary cleanup once the go routine is cancelled.

We also change the assumption that Run() can be called more than once;
now you must call Run() only once and invoking multiple times has
undefined behavior (and will most likely result in a panic).

The reason for this change is that, in practice, it is very easy to
ensure that Run() is only called once and the code to track the state of
a service (if it is running or not) adds unnecessary overhead.

This patch also adds a services package which we use to coordinate and
share information between services.