fix: Allow excessive precision for now
The way these tests are written, any truncation will be made on both the
values used in the test and the values expected in the test, so this is
probably fine.

But we might not be testing with the exact values we think we're testing
with, so we should audit the values clippy warns about at some point.