Add Partition Store
This adds the Partition trait along with helper functions and structs to combine the results from multiple partitions.

It also refactors MemDB to have methods that align with the partition trait so that it can be used later by database.rs.

There are obvious places where it is inefficient, but this should be enough to get started with refactoring database.rs to use this rather than rocksdb, series_store, and inverted_index.