chore(pkger): refactor influx pkg cmd to not validate in CLI and have server validate
the pkger.ValidSkipParseError option allows our server to be the one to validate the
the pkg is accurate. If a user has an older version of the UI and our cloud gets updated
with new validation rules,they'll get immediate access to that change without having to
rol their CLI build.

also fixes issue where we swallow initial errors when check setup middleware
fails.