fix(snapshotter): properly read payload
This is a fix and a bit of a refactor of Service.readRequest().

This commit changes the signature so that it accepts an io.Reader
instead of a net.Conn because there's nothing about the function that
relies on it being a Conn; only a Reader.

We also changed the signature so that it returns a *Request instead of
Request -- this way we don't allocate an entire Request value on error.

I also tried to document the few edge cases that the function has to
account for and one quirk where we expect to read a newline which isn't
accounted for by Request.UploadSize.