fix(influx): make CLI respect root token and host env vars in additionn to config
one thing to note here is we are deleting the default value on the host
flag when it is registered. The config is the fallback and has the default
value set. If the host flag has a default, the determination if the user
set it or not is ambiguous. We can't have that.

closes: #17812