fix(query): Detach killed queries
Before this commit, killing a query leaves it in a "killed" state such
that it is visible in the tasks list after being killed.

Sometimes tasks aren't properly removed from the list. It seems like a
better approach would be to kill and remove a task from the list when
the KILL QUERY command is issued.

This commit wraps the TaskManager.DetachDetach() method with query.Kill()
-- this way, when we call TaskManager.KillQuery(), it properly kills and
detaches the task.