fix: need to rebuild the query request before the second e2e test run
The e2e tests in influxdb run the tests twice to get past the fact that there
is no way to force order between the write to storage and the read back. When
the json.Marshal call became mandatory it was added to the first run, but not
the second.